refactor: useLogout user hook으로 디렉터리 위치 변경
